requirements.txt
boto3==1.9.210
botocore==1.12.210
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.2
django-decouple==2.1
django-storages==1.7.1
djangorestframework==3.9.4
docutils==0.15.2
gunicorn==19.9.0
idna==2.8
jmespath==0.9.4
oauthlib==3.0.1
PyJWT==1.7.1
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
s3transfer==0.2.1
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
urllib3==1.25.3
psycopg2>=2.6.1
